Changed mv calls to cp and cp -r, and added a containerfile.
The mv calls are causing issues on shared filesystems. See reports:

- https://github.com/dcouvin/CRISPRCasFinder/issues/35
- https://github.com/EBI-Metagenomics/mobilome-annotation-pipeline/issues/37

This patch is not an ideal solution; the proper approach would be to avoid moving or copying files and instead
generate them directly in their final folder. However, I don't have the capacity to address this or the necessary Perl knowledge.

The container is compatible with Docker and Singularity.